fix api missing fullname and improve login.html page
