--- BA fb/PhishingFacebook.pem
@@ -1,27 +1,0 @@
------BEGIN RSA PRIVATE KEY-----
-MIIEpQIBAAKCAQEAo9TrbU/bwjJNHT6wKMC1pxa0pHxEJK4yldhB5dnfRUTQPLT2
-K/hoUvfHTGz2fO+muNAO8Y1KRzc6/2rEZq8k5ujDyltUxCXa/j5EWt0PYgKIk1Hf
-XKIUQVZEad+xQTrImguov2PPic7SuvukXP/E9A+DtW2/HvnNDJkhC0Gqk/6ED598
-ReRH2r82iCVcdrbNjcjrQDQv8UbN5ilA7osk31WiEN28PTNakOZPI87QyjQeFM+w
-o9tdPtDKj3hyJRNtJJ8l5e6S2WfGOXTcRQFXXGUVtsc4rZ/RtjvPjCHWzCd3Eoes
-CrHXoD3I3VgR7vlaRzJ87wdrrRTgbUPybtB2TwIDAQABAoIBAB2IdmAgMzdQpgIN
-mOffxV83gpzcIdSS6l+Eumxa2Bq+iuNGo9HNV9ENBYpLYiunb+CE8QyuMaOg2yvL
-7VetXvOZJiJGUIODZ3TJeL0AeJNsbX5RFfi25vvYN0D2qBdgJ2VNkL/fmDv5Ypd+
-YWS6XqC4Yz/DZt6InWQV/pTu43cF8qKcyDPWqBcTsllQ7USfF+EiqnbH0N912Yf8
-IvRlYDF4UTU3s41P6OCxR5VmF1ws0QSjrojaSDFnDqGGEXxVUmVTYWdjrhgfgEIq
-bgQvAARo9JzdhBEv4yNshrHs/dujuJHy3HE3m3tatFFs/YaKShHbqMBnX8zXMzUJ
-TbmQujECgYEA0QMzOUhaopxxfrF6GIYoVQKgSxavLiYYSx4pR/QBGUMS7GuBLCPM
-WBrHeQxra+HJ7WaWv+v9zvpMVnGXvk6yw/6afEHCQiN2Xfcj9nJk/1SC0a/Tultl
-mKE/RxKUiYfu2vwoLEJna7YcJhSJOazGmHuAnE1oW9VYhgesDkIBjMUCgYEAyKmJ
-inElXKhFR5HzMcppMGFTR4D4bLmoRTNxaYRZJZhZXGD/SC3IeT0rnXoIlXDtTmwo
-1hjLuJKhJJQAy1yJKpYYoDhd8u4/lNyK4W6mew+ezxqfvlEaA/bJbz3360PA54v2
-ZSCoGqJb48oJVLyXB2Tb0NrTR76fAxCYe0b5kAMCgYEAg3xL4SFPQzS/gUAM2xiq
-jrAUPqiKk7GLXtYeihFKD6ZOZURMAoSSrGpMSi4x9bAY46vB8gJ3rJlxJIBZQXMo
-qnukm76SgiDk+b/bc9ILD0UO9vJxc67Dfv0RXZimsM4+yKwQaadsgCoUEzCgExRJ
-6urmwW1A1WSlR8vTxTRo3TECgYEAqUmuyedu997Wo+zT2XvEhpZ1Zeu+qitvdQJT
-egu4EemX6kHC3SgS64XZ1/9uYnVlHzZMtWMx68N4w51Blk+5QP5kRDeJuzuY1IPD
-fzCUr1mcR2lUv4E4i4OWqY0ELa15f18rPmHwsJetIl3+xUhdulUyxd73RJ3jFPjZ
-S5ezT/UCgYEAlMty6FmybEr8sHUUgWLFS6dNNLQlKnBOoMAAb2mvxMDtZE782cVj
-zALke+2y1MmuGwDAlx2ZKk4J3Qksu/nc685JJaSaT/O1XFiBcRYig8JeOtKGrUGS
-298DMLRsd3oWn4me+fZUzoG/R/QqXDij6Ks32kIf5OcC41WWzQQFbT4=
------END RSA PRIVATE KEY-----
